requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
boto3==1.21.37
botocore==1.24.46
certifi==2023.7.22
charset-normalizer==3.3.2
graypy==2.1.0
greenlet==3.0.1
idna==3.4
jmespath==1.0.1
numpy==1.26.2
pandas==2.1.3
pika==1.3.2
playwright==1.40.0
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
pyee==11.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.5.2
six==1.16.0
soupsieve==2.5
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
